refact structure directory, add cart module, move checkout to helpers dir
