+++ src/dom/listeners/addToCartListener.js
@@ -1,8 +1,8 @@
 import { addToCartButtons } from '@dom/selectors'
-import addToCart from '@modules/addToCart'
-import calculateTotal from '@modules/calculateTotal'
-import counterCart from '@modules/counterCart'
-import applyPromotionsCart from '@modules/applyPromotionsCart'
+import addToCart from '@modules/cart/addToCart'
+import calculateTotal from '@modules/cart/calculateTotal'
+import counterCart from '@modules/cart/counterCart'
+import applyPromotionsCart from '@modules/cart/applyPromotionsCart'
 
 const addCartListener = () => {
     let idProduct = 0
