+++ src/dom/listeners/modalListener.js
@@ -1,6 +1,6 @@
 import { launchShoppingCard, cleanCartButton } from '@dom/selectors'
-import cleanCart from '@modules/cleanCart'
-import openModal from '@modules/openModal'
+import cleanCart from '@modules/cart/cleanCart'
+import openModal from '@modules/cart/openModal'
 
 const modalListener = () => {
     if (!launchShoppingCard || !cleanCartButton) return
