+++ src/modules/cart/cleanCart.js
@@ -1,4 +1,5 @@
 import cart from '@content/cart.js'
+import counterCart from '@modules/cart/counterCart.js'
 
 const cleanCart = () => {
     const shoppingCardModalBoy = document.querySelector(
@@ -6,5 +7,6 @@ const cleanCart = () => {
     )
     cart.length = 0
     shoppingCardModalBoy.classList.add('d-none')
+    counterCart()
 }
 export default cleanCart
